Update del backend
