+++ Dockerfile
@@ -1,4 +1,8 @@
 FROM tiangolo/uvicorn-gunicorn-fastapi
-RUN pip install uvicorn
-EXPOSE 80
-COPY . /app
+RUN pip install peewee
+RUN pip install aiofiles
+RUN pip install pandas
+RUN pip install jinja2==3.0.3
+RUN pip install psycopg2-binary
+RUN pip install python-multipart
+COPY ./app /app
