--- static/styles.css
@@ -1,18 +1,0 @@
-.container {
-    font-family: Arial, Helvetica, sans-serif;
-    color:steelblue;
-    text-align: center;
-    width: 100%;
-}
-
-.form {
-    font-family: Arial, Helvetica, sans-serif;
-    text-align: center;
-    width: 100%;
-}
-
-.form1 {
-    font-family: Arial, Helvetica, sans-serif;
-    text-align: center;
-    width: 30%;
-}
