--- templates/index.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Formulario SUP</title>
-    <link href="{{ url_for('static', path='/styles.css') }}" rel="stylesheet">
-</head>
-<body>
-    <div class="container">
-        <h1>{{message}}</h1>
-    </div>
-    <div class="form">
-        <form action="/add_contact", method="POST">
-
-            <input type="text" name="nombre" placeholder="Nombre" class="form1"> 
-            <br>
-            <br>
-            <input type="text" name="apellido" placeholder="Apellido" class="form1">
-            <br>
-            <br>
-            <input type="text" name="edad" placeholder="Edad" class="form1">
-            <br>
-            <br>
-            <input type="text" name="email" placeholder="Email" class="form1">
-            <br>
-            <br>
-            <button type="submit">
-                Enviar
-            </button>
-        
-        </form>
-    </div>
-</body>
-</html>
